ClientModeBox.py: download EPG y Channels in USB HDD and Internal Flash

+++ src/ClientModeBox.py
@@ -654,6 +654,8 @@ class ClientModeBoxMenu(Screen, ConfigListScreen):
 
 class ClientModeBoxDownloader:
 	DIR_ENIGMA2 = "/etc/enigma2/"
+	DIR_HDD = "/media/hdd/"
+	DIR_USB = "/media/usb/"
 	DIR_TMP = "/tmp/"
 
 	def __init__(self, session):
@@ -821,7 +823,7 @@ class ClientModeBoxDownloader:
 		db.reloadBouquets()
 		AddNotificationWithID("ClientModeBoxChannelsImportOK", MessageBox, _("Channels imported successfully from %s") % baseurl, type=MessageBox.TYPE_INFO, timeout=5)
 
-	def downloadEPGFTP(self, baseurl): # PYTHON 3
+	def downloadEPGFTP(self, baseurl):
 		print("[ClientModeBox] downloadEPGFTP Force EPG save on remote receiver...")
 		self.forceSaveEPGonRemoteReceiver(baseurl)
 		print("[ClientModeBox] downloadEPGFTP Searching for epg.dat...")
@@ -877,15 +879,42 @@ class ClientModeBoxDownloader:
 	def checkEPGCallback(self):
 		baseurl = "http://"
 		baseurl += config.ipboxclient.host.value
-		self.remoteEPGpath = self.DIR_ENIGMA2
-		self.remoteEPGfile = "epg"
-		self.remoteEPGfile = "%s.dat" % self.remoteEPGfile.replace('.dat', '')
-		print("[ClientModeBox] Remote EPG filename. '%s%s'" % (self.remoteEPGpath, self.remoteEPGfile))
-		result = self.FTPdownloadFile(self.remoteEPGpath, self.remoteEPGfile, "epg.dat")
-		if result:
-			self.importEPGCallback()
-		else:
-			AddNotificationWithID("EPGImportNOK", MessageBox, _("EPG not imported check that path EPG in %s is in internal flash") % baseurl, type=MessageBox.TYPE_ERROR, timeout=5)
+		try:
+			self.remoteEPGpath = self.DIR_ENIGMA2
+			self.remoteEPGfile = "epg"
+			self.remoteEPGfile = "%s.dat" % self.remoteEPGfile.replace('.dat', '')
+			print("[ClientModeBox] Remote EPG filename. '%s%s'" % (self.remoteEPGpath, self.remoteEPGfile))
+			result = self.FTPdownloadFile(self.remoteEPGpath, self.remoteEPGfile, "epg.dat")
+			if result:
+				self.importEPGCallback()
+			else:
+				print("[ClientModeBox] Remote EPG filename not path in internal flash")
+		except Exception as err:
+			print("[ClientModeBox] cannot save EPG %s" % err)
+		try:
+			self.remoteEPGpath = self.DIR_HDD
+			self.remoteEPGfile = "epg"
+			self.remoteEPGfile = "%s.dat" % self.remoteEPGfile.replace('.dat', '')
+			print("[ClientModeBox] Remote EPG filename. '%s%s'" % (self.remoteEPGpath, self.remoteEPGfile))
+			result = self.FTPdownloadFile(self.remoteEPGpath, self.remoteEPGfile, "epg.dat")
+			if result:
+				self.importEPGCallback()
+			else:
+				print("[ClientModeBox] Remote EPG filename not path in HDD")
+		except Exception as err:
+			print("[ClientModeBox] cannot save EPG %s" % err)
+		try:
+			self.remoteEPGpath = self.DIR_USB
+			self.remoteEPGfile = "epg"
+			self.remoteEPGfile = "%s.dat" % self.remoteEPGfile.replace('.dat', '')
+			print("[ClientModeBox] Remote EPG filename. '%s%s'" % (self.remoteEPGpath, self.remoteEPGfile))
+			result = self.FTPdownloadFile(self.remoteEPGpath, self.remoteEPGfile, "epg.dat")
+			if result:
+				self.importEPGCallback()
+			else:
+				print("[ClientModeBox] Remote EPG filename not path in USB")
+		except Exception as err:
+			print("[ClientModeBox] cannot save EPG %s" % err)
 
 	def copyFile(self, source, dest):
 		shutil.copy2(source, dest)
@@ -893,7 +922,7 @@ class ClientModeBoxDownloader:
 	def importEPGCallback(self):
 		baseurl = "http://"
 		baseurl += config.ipboxclient.host.value
-		print("[ClientModeBox] importEPGCallback '%s%s' downloaded successfully. " % (self.remoteEPGpath, self.remoteEPGfile))
+		print("[ClientModeBox] importEPGCallback '%s%s' downloaded successfully from server." % (self.remoteEPGpath, self.remoteEPGfile))
 		print("[ClientModeBox] importEPGCallback Removing current EPG data...")
 		try:
 			os.remove(config.misc.epgcache_filename.value)
@@ -904,7 +933,7 @@ class ClientModeBoxDownloader:
 		eEPGCache.getInstance().load()
 		print("[ClientModeBox] importEPGCallback New EPG data loaded...")
 		print("[ClientModeBox] importEPGCallback Closing importer.")
-		AddNotificationWithID("EPGImportOK", MessageBox, _("EPG imported successfully from %s") % baseurl, type=MessageBox.TYPE_INFO, timeout=5)
+		AddNotificationWithID("ClientModeBoxEPGImportOK", MessageBox, _("EPG imported successfully from %s") % baseurl, type=MessageBox.TYPE_INFO, timeout=5)
 
 	def downloadEPG(self, baseurl):
 		from json import loads
@@ -914,11 +943,11 @@ class ClientModeBoxDownloader:
 			print("[ClientModeBox] %s" % err)
 		print("[ClientModeBox] Get EPG Location")
 		try:
-			epgdatfile = "/media/hdd/epg.dat" or "/media/usb/epg.dat" or "/etc/enigma2/epg.dat"
+			epgdatfile = "/etc/enigma2/epg.dat"
 			files = [file for file in loads(urlopen("%s/file?dir=%s" % (baseurl, os.path.dirname(epgdatfile)), timeout=5).read())["files"] if os.path.basename(file).startswith("epg.dat")]
 			epg_location = files[0] if files else None
 			if epg_location:
-				print("[ClientModeBox] Copy EPG file...")
+				print("[ClientModeBox] Copy EPG file from %s" % baseurl)
 				try:
 					try:
 						os.mkdir("/tmp/epgdat")
@@ -927,15 +956,64 @@ class ClientModeBoxDownloader:
 					epgdattmp = "/tmp/epgdat"
 					epgdatserver = "/tmp/epgdat/epg.dat"
 					open("%s/%s" % (epgdattmp, os.path.basename(epg_location)), "wb").write(urlopen("%s/file?file=%s" % (baseurl, epg_location), timeout=5).read())
-					shutil.move("%s" % epgdatserver, "%s" % (config.misc.epgcache_filename.value))
-					eEPGCache.getInstance().load()
-					shutil.rmtree(epgdattmp)
-					AddNotificationWithID("EPGImportOK", MessageBox, _("EPG imported successfully from %s") % baseurl, type=MessageBox.TYPE_INFO, timeout=5)
+					if "epg.dat" in (epgdatserver):
+						shutil.move("%s" % epgdatserver, "%s" % (config.misc.epgcache_filename.value))
+						eEPGCache.getInstance().load()
+						shutil.rmtree(epgdattmp)
+						AddNotificationWithID("ClientModeBoxEPGImportOK", MessageBox, _("EPG imported successfully from %s") % baseurl, type=MessageBox.TYPE_INFO, timeout=5)
 				except Exception as err:
 					print("[ClientModeBox] cannot save EPG %s" % err)
 		except Exception as err:
 			print("[ClientModeBox] cannot save EPG %s" % err)
-			self.downloadEPGFTP(baseurl)
+			return self.downloadEPGFTP(baseurl)
+		try:
+			epgdatfile = "/media/hdd/epg.dat"
+			files = [file for file in loads(urlopen("%s/file?dir=%s" % (baseurl, os.path.dirname(epgdatfile)), timeout=5).read())["files"] if os.path.basename(file).startswith("epg.dat")]
+			epg_location = files[0] if files else None
+			if epg_location:
+				print("[ClientModeBox] Copy EPG file from %s" % baseurl)
+				try:
+					try:
+						os.mkdir("/tmp/epgdat")
+					except:
+						print("[ClientModeBox] epgdat folder exists in tmp")
+					epgdattmp = "/tmp/epgdat"
+					epgdatserver = "/tmp/epgdat/epg.dat"
+					open("%s/%s" % (epgdattmp, os.path.basename(epg_location)), "wb").write(urlopen("%s/file?file=%s" % (baseurl, epg_location), timeout=5).read())
+					if "epg.dat" in (epgdatserver):
+						shutil.move("%s" % epgdatserver, "%s" % (config.misc.epgcache_filename.value))
+						eEPGCache.getInstance().load()
+						shutil.rmtree(epgdattmp)
+						AddNotificationWithID("ClientModeBoxEPGImportOK", MessageBox, _("EPG imported successfully from %s") % baseurl, type=MessageBox.TYPE_INFO, timeout=5)
+				except Exception as err:
+					print("[ClientModeBox] cannot save EPG %s" % err)
+		except Exception as err:
+			print("[ClientModeBox] cannot save EPG %s" % err)
+			return self.downloadEPGFTP(baseurl)
+		try:
+			epgdatfile = "/media/usb/epg.dat"
+			files = [file for file in loads(urlopen("%s/file?dir=%s" % (baseurl, os.path.dirname(epgdatfile)), timeout=5).read())["files"] if os.path.basename(file).startswith("epg.dat")]
+			epg_location = files[0] if files else None
+			if epg_location:
+				print("[ClientModeBox] Copy EPG file from %s" % baseurl)
+				try:
+					try:
+						os.mkdir("/tmp/epgdat")
+					except:
+						print("[ClientModeBox] epgdat folder exists in tmp")
+					epgdattmp = "/tmp/epgdat"
+					epgdatserver = "/tmp/epgdat/epg.dat"
+					open("%s/%s" % (epgdattmp, os.path.basename(epg_location)), "wb").write(urlopen("%s/file?file=%s" % (baseurl, epg_location), timeout=5).read())
+					if "epg.dat" in (epgdatserver):
+						shutil.move("%s" % epgdatserver, "%s" % (config.misc.epgcache_filename.value))
+						eEPGCache.getInstance().load()
+						shutil.rmtree(epgdattmp)
+						AddNotificationWithID("ClientModeBoxEPGImportOK", MessageBox, _("EPG imported successfully from %s") % baseurl, type=MessageBox.TYPE_INFO, timeout=5)
+				except Exception as err:
+					print("[ClientModeBox] cannot save EPG %s" % err)
+		except Exception as err:
+			print("[ClientModeBox] cannot save EPG %s" % err)
+			return self.downloadEPGFTP(baseurl)
 
 	def downloadParentalControl(self, baseurl):
 		print("[ClientModeBox] reading remote parental control status ...")
